litevlm/text_decoder.py
"""
Text Decoder Module
Qwen2 기반 텍스트 디코더 (TensorRT 최적화)
"""

from pathlib import Path
from typing import Union, Optional

import torch
import torch.nn as nn


class TextDecoder:
    """
    Text Decoder (Qwen2 기반)
    
    Features:
    - TensorRT 엔진 지원
    - FP8/INT8 양자화 지원
    - KV-cache 최적화
    """
    
    def __init__(
        self,
        model_path: Union[str, Path],
        device: str = "cuda",
        use_kv_cache: bool = True,
    ):
        """
        Args:
            model_path: TensorRT 엔진 파일 또는 PyTorch 모델 경로
            device: 실행 디바이스
            use_kv_cache: KV-cache 사용 여부
        """
        self.device = device
        self.use_kv_cache = use_kv_cache
        
        model_path = Path(model_path)
        
        # TensorRT 엔진 로드
        if model_path.suffix == ".engine":
            self.model = self._load_tensorrt_engine(model_path)
            self.use_tensorrt = True
        else:
            # PyTorch 모델 로드
            self.model = self._load_pytorch_model(model_path)
            self.use_tensorrt = False
            self.model.to(device)
            self.model.eval()
    
    def _load_tensorrt_engine(self, engine_path: Path):
        """TensorRT 엔진 로드"""
        try:
            import tensorrt as trt
            import pycuda.driver as cuda
            import pycuda.autoinit
            
            TRT_LOGGER = trt.Logger(trt.Logger.WARNING)
            
            with open(engine_path, "rb") as f:
                engine_data = f.read()
            
            runtime = trt.Runtime(TRT_LOGGER)
            engine = runtime.deserialize_cuda_engine(engine_data)
            
            return engine
        
        except ImportError:
            raise ImportError(
                "TensorRT not installed. Please install TensorRT or use PyTorch model."
            )
    
    def _load_pytorch_model(self, model_path: Path):
        """PyTorch 모델 로드"""
        try:
            from transformers import AutoModelForCausalLM
            
            # Qwen2 모델 로드
            model = AutoModelForCausalLM.from_pretrained(
                model_path,
                trust_remote_code=True,
                torch_dtype=torch.float16,
            )
            
            return model
        
        except Exception as e:
            print(f"[Warning] Could not load Qwen2 model: {e}")
            print("[Warning] Using dummy decoder")
            return DummyDecoder()
    
    def generate(
        self,
        input_embeds: torch.Tensor,
        max_new_tokens: int = 256,
        temperature: float = 0.7,
        top_p: float = 0.9,
        top_k: int = 50,
        repetition_penalty: float = 1.1,
    ) -> torch.Tensor:
        """
        텍스트 생성
        
        Args:
            input_embeds: 입력 임베딩 [B, L, D]
            max_new_tokens: 최대 생성 토큰 수
            temperature: 샘플링 온도
            top_p: Nucleus sampling
            top_k: Top-K sampling
            repetition_penalty: 반복 페널티
            
        Returns:
            output_ids: 생성된 토큰 IDs [B, L']
        """
        if self.use_tensorrt:
            return self._generate_tensorrt(
                input_embeds,
                max_new_tokens,
                temperature,
                top_p,
                top_k,
                repetition_penalty
            )
        else:
            return self._generate_pytorch(
                input_embeds,
                max_new_tokens,
                temperature,
                top_p,
                top_k,
                repetition_penalty
            )
    
    def _generate_pytorch(
        self,
        input_embeds: torch.Tensor,
        max_new_tokens: int,
        temperature: float,
        top_p: float,
        top_k: int,
        repetition_penalty: float,
    ) -> torch.Tensor:
        """PyTorch 모델로 생성"""
        with torch.no_grad():
            # Hugging Face generate 사용
            if hasattr(self.model, 'generate'):
                outputs = self.model.generate(
                    inputs_embeds=input_embeds,
                    max_new_tokens=max_new_tokens,
                    temperature=temperature,
                    top_p=top_p,
                    top_k=top_k,
                    repetition_penalty=repetition_penalty,
                    do_sample=temperature > 0,
                    use_cache=self.use_kv_cache,
                )
                return outputs
            
            # 간단한 autoregressive 생성
            return self._simple_generate(
                input_embeds,
                max_new_tokens,
                temperature
            )
    
    def _generate_tensorrt(
        self,
        input_embeds: torch.Tensor,
        max_new_tokens: int,
        temperature: float,
        top_p: float,
        top_k: int,
        repetition_penalty: float,
    ) -> torch.Tensor:
        """TensorRT 엔진으로 생성"""
        # TensorRT 추론 로직
        raise NotImplementedError("TensorRT generation to be implemented")
    
    def _simple_generate(
        self,
        input_embeds: torch.Tensor,
        max_new_tokens: int,
        temperature: float
    ) -> torch.Tensor:
        """간단한 autoregressive 생성"""
        B, L, D = input_embeds.shape
        
        # 더미 토큰 ID 반환
        # 실제 구현에서는 autoregressive decoding 수행
        output_ids = torch.randint(0, 32000, (B, max_new_tokens), device=self.device)
        
        return output_ids
    
    def __call__(self, *args, **kwargs):
        """Forward 호출"""
        return self.generate(*args, **kwargs)


class DummyDecoder(nn.Module):
    """더미 디코더 (테스트용)"""
    
    def __init__(self):
        super().__init__()
        self.dummy = nn.Linear(1, 1)
    
    def forward(self, input_embeds):
        return input_embeds
    
    def generate(self, inputs_embeds, **kwargs):
        B = inputs_embeds.shape[0]
        max_new_tokens = kwargs.get('max_new_tokens', 50)
        return torch.randint(0, 32000, (B, max_new_tokens))
