requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
optimum>=1.14.0

# Vision & Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
timm>=0.9.0

# Model Optimization
# tensorrt>=8.5.0  # Included in JetPack
onnx>=1.14.0
onnxruntime>=1.16.0
# pycuda>=2022.1  # Included in JetPack

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
einops>=0.7.0
safetensors>=0.4.0

# CLI & Configuration
click>=8.1.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Monitoring & Profiling
psutil>=5.9.0
pynvml>=11.5.0

# HuggingFace Hub
huggingface-hub>=0.19.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Optional: For development
pytest>=7.4.0
black>=23.10.0
flake8>=6.1.0
ipython>=8.12.0
